refactor: move 관련 메서드 ChessGame으로 분리
